io/kqueue: extract shared substrate into kqueue_common.{h,c}
Commit 4a of the macOS backend series (commit 4 was split into
4a/4b per the plan addendum W-5 — 3-way extraction, not 2).
Preparatory move of the kqueue substrate shared by both the
FreeBSD aio backend (backend_kqueue.c) and the forthcoming
Darwin thread-pool backend (backend_darwin.c).

New files:

  lib/io/kqueue_common.h: struct ring_context definition
    (was static in backend_kqueue.c), declarations for the
    substrate primitives.

  lib/io/kqueue_common.c: definitions of ring_context_alloc/free,
    kq_setup / kq_teardown (kqueue fd + shutdown pipe + EVFILT_READ
    registration), io_backend_set/get_global,
    io_network_set/get_global (new setter symmetric with the
    existing getters -- io_handler_init now uses it instead of
    poking g_network_rc directly), io_handler_stop,
    io_handler_signal_shutdown.

Moved out of backend_kqueue.c:

  - struct ring_context (kqueue variant).
  - g_backend_rc / g_network_rc globals.
  - io_backend_set_global, io_backend_get_global,
    io_network_get_global accessors.
  - ring_context_alloc, ring_context_free.
  - kq_setup (now non-static), kq_teardown (now non-static).
  - io_handler_stop (was inline empty; new impl is documented
    no-op since the shutdown flag is per-loop volatile).
  - io_handler_signal_shutdown.

backend_kqueue.c now includes kqueue_common.h for these
symbols.  io_handler_init at line 377 calls io_network_set_global
instead of assigning directly to the global.

Makefile.am: kqueue_common.c compiles under IO_BACKEND_KQUEUE AND
(when the Darwin gate flips on in commit 7) IO_BACKEND_DARWIN.
Listed in both conditionals.

No Linux build change (backend.c still has its own getters;
handler.c still has io_network_get_global).  No FreeBSD runtime
change; pure code move across files.

Splitting the socket-side extraction (io_handler_init/fini,
io_request_accept_op, kqueue_request_connect, read/write/accept
dispatch, heartbeat, io_send_request, etc.) to kqueue_socket.c
is commit 4b.
